requirements.txt
streamlit
langgraph
groq
python-dotenv
duckduckgo-search
pydantic